+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @richie-router/server
 
+## 0.1.9
+
+### Patch Changes
+
+- add support for @ in urls
+- Updated dependencies
+  - @richie-router/core@0.1.6
+
 ## 0.1.8
 
 ### Patch Changes